changed R70 again

+++ src/ESFA.DC.ILR.TestDataGenerator.Functors/CrossEntity/R70.cs
@@ -39,10 +39,10 @@ namespace DCT.TestDataGenerator.Functor
 
         private void MutateCommon(MessageLearner learner, bool valid)
         {
-            int aim;
-            learner.LearningDelivery[0].AimType = AimType.ComponentAim;
-            learner.LearningDelivery[0].ProgType = ProgType.ApprenticeshipStandard;
-            learner.LearningDelivery[1].ProgType = ProgType.ApprenticeshipStandard;
+            long aim;
+            learner.LearningDelivery[0].AimType = (long)AimType.ComponentAim;
+            learner.LearningDelivery[0].ProgType = (long)ProgType.ApprenticeshipStandard;
+            learner.LearningDelivery[1].ProgType = (long)ProgType.ApprenticeshipStandard;
             aim = valid ? 1 : 2;
             learner.LearningDelivery[1].AimType = aim;
         }
